Separates constraint solver logic
