+++ src/analysis/interproc/digstack.rs
@@ -288,7 +288,7 @@ mod test {
         let mut ssa = SSAStorage::new();
         {
             let mut constructor = SSAConstruct::new(&mut ssa, &reg_profile);
-            constructor.run(instructions.ops.unwrap());
+            constructor.run(instructions.ops.unwrap().as_slice());
         }
         {
             dce::collect(&mut ssa);
@@ -313,7 +313,7 @@ mod test {
         let mut ssa = SSAStorage::new();
         {
             let mut constructor = SSAConstruct::new(&mut ssa, &reg_profile);
-            constructor.run(instructions.ops.unwrap());
+            constructor.run(instructions.ops.unwrap().as_slice());
         }
         {
             dce::collect(&mut ssa);
