+++ src/analysis/mod.rs
@@ -26,3 +26,4 @@ pub mod matcher {
 pub mod interproc;
 pub mod tie;
 pub mod mark_refs;
+pub mod constraint_set;
