+++ tests/bin_ls_test.rs
@@ -42,7 +42,7 @@ fn run_construction() -> SSAStorage {
     let mut ssa = SSAStorage::new();
     {
         let mut constructor = SSAConstruct::new(&mut ssa, &reg_profile);
-        constructor.run(instructions.ops.unwrap());
+        constructor.run(instructions.ops.unwrap().as_slice());
     }
     {
         dce::collect(&mut ssa);
